-added migrations -update ignores
