+++ .gitignore
@@ -8,7 +8,6 @@ dist
 /.pydevproject
 /.project
 /.settings
-dataforms/migrations
 example/static/*
 example/media/*
 example/dataforms.db
