+++ dataforms/models.py
@@ -119,7 +119,9 @@ class Field(models.Model):
     slug = models.SlugField(verbose_name=_('slug'), max_length=255, unique=True, validators=[reserved_delimiter])
     help_text = models.TextField(verbose_name=_('field help text'), blank=True)
     initial = models.TextField(verbose_name=_('initial value of the field'), blank=True)
-    arguments = models.CharField(verbose_name=_('additional arguments'),
+    classes = models.CharField(verbose_name=_('additional widget classes'), 
+                                      help_text="A comma separated string of class names.", blank=True, max_length=255)
+    arguments = models.CharField(verbose_name=_('additional field arguments'),
         help_text="A JSON dictionary of keyword arguments.", blank=True, max_length=255)
     required = models.BooleanField(verbose_name=_('field is required'), default=False)
     visible = models.BooleanField(verbose_name=_('field is visible'), default=True)
