+++ dataforms/templates/dataforms/base.html
@@ -1,0 +1,2 @@
+{% extends "admin/base_site.html" %}
+
