+++ dataforms/templates/dataforms/get_field.html
@@ -1,0 +1,5 @@
+<li>
+	<img src="{{ STATIC_URL }}dataforms/img/arrow.png" alt="arrow" />
+	<label for="{{ title }}">{{ title }}</label>
+	{{ field }}
+</li>
